notification_service/requirements.txt
# Python dependencies
firebase-admin>=5.0.0
APScheduler>=3.9.1
psycopg2-binary>=2.9.1
pytz>=2021.3
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0